requirements.txt
streamlit==1.33.0
Pillow==10.3.0
easyocr==1.7.1
pandas==2.2.1
numpy==1.26.4